+++ CHANGELOG.markdown
@@ -1,7 +1,7 @@
 next
 ----
-* Generalized types of `transformMOf`, transformOf`, `transformMOnOf`,
-  `transformOnOf`, `rewriteMOf`, rewriteOf`, `rewriteMOnOf` and `rewriteOnOf`.
+* Generalized types of `transformMOf`, `transformOf`, `transformMOnOf`,
+  `transformOnOf`, `rewriteMOf`, `rewriteOf`, `rewriteMOnOf` and `rewriteOnOf`.
 * Depend on `th-abstraction` package for normalizing differences across
   `template-haskell` versions